Added SPDX license identifiers and copyright notices
In this update, we've added SPDX license identifiers and copyright notices to all source files. We also created a new DEP5 file for reuse compliance. The AGPL-3.0-or-later license text has been added to the project's LICENSES directory. Changes have been made in the LICENSE.md file as well to reflect these updates.